HTTP Handler 상위 Manager 클래스 정의 및 상속 구조 설계, BarBaseActor의 value 텍스트렌더 컴포넌트의 z축 조정 오프셋 계산 로직 추정

+++ Source/test/Public/Datas/JCMBarBaseActor.h
@@ -35,7 +35,7 @@ private:
 	int32 SpawnCount = 0;
 
 	UPROPERTY()
-	float SpawnedCustomMeshAmount;
+	int32 SpawnedCustomMeshAmount;
 
 	// TimeLine Animation Binding Function
 	UFUNCTION()
